⚡️ day9 part2 : improve performances

+++ adventOfCode/src/main/kotlin/AdventOfCode2018/Solutions.kt
@@ -13,7 +13,7 @@ class Solutions {
     var day1 = Day<Long, Long>()
     var day2 = Day<Int, String>()
     var day8 = Day<Int, Int>()
-    var day9 = Day<Int, Int>()
+    var day9 = Day<Long, Long>()
 
     companion object {
         fun readFile():Solutions{
